Fake data for repositories. Small things

+++ src/OzonEdu.MerchandiseService.Domain/AggregationModels/ManagerAggregate/Manager.cs
@@ -32,7 +32,7 @@ namespace OzonEdu.MerchandiseService.Domain.AggregationModels.ManagerAggregate
 
         public void AssignTask() => AssignedTasks++;
 
-        public Manager SetId(int id)
+        public Manager SetId(long id)
         {
             Id = id;
             return this;
